requirements.txt
alpaca-py>=0.21.0
apscheduler>=3.10.0
flask>=3.0.0
gunicorn>=21.2.0
pandas>=2.1.0
ta>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
